+++ src/modules/palette/private/GimpPalette.h
@@ -11,6 +11,8 @@ namespace palette {
 /**
  * Gimp format with Aseprite extension for alpha channels
  *
+ * https://github.com/LibreSprite/LibreSprite/tree/master/data/palettes
+ *
  * @ingroup Formats
  */
 class GimpPalette : public PaletteFormat {