+++ src/modules/voxelformat/VolumeFormat.cpp
@@ -19,6 +19,7 @@
 #include "scenegraph/SceneGraphNode.h"
 #include "video/Texture.h"
 #include "voxelformat/Format.h"
+#include "voxelformat/private/aseprite/AsepriteFormat.h"
 #include "voxelformat/private/rooms/ThingFormat.h"
 #include "voxelformat/private/aceofspades/AoSVXLFormat.h"
 #include "voxelformat/private/animatoon/AnimaToonFormat.h"
@@ -100,6 +101,14 @@ const io::FormatDescription &vengi() {
 								   {"vengi"},
 								   {"VENG"},
 								   VOX_FORMAT_FLAG_PALETTE_EMBEDDED | VOX_FORMAT_FLAG_ANIMATION | FORMAT_FLAG_SAVE};
+	return f;
+}
+
+static const io::FormatDescription &aseprite() {
+	static io::FormatDescription f{"aseprite",
+								   {"aseprite", "ase"},
+								   {},
+								   VOX_FORMAT_FLAG_PALETTE_EMBEDDED};
 	return f;
 }
 
@@ -335,6 +344,7 @@ const io::FormatDescription *voxelLoad() {
 	// this is the list of supported voxel volume formats that are
 	// have importers implemented
 	static const io::FormatDescription desc[] = {vengi(),
+													aseprite(),
 												 qubicleBinary(),
 												 magicaVoxel(),
 												 qubicleBinaryTree(),
@@ -479,6 +489,8 @@ static core::SharedPtr<Format> getFormat(const io::FormatDescription &desc, uint
 			return core::make_shared<CubzhFormat>();
 		} else if (ext == cubzhB64().mainExtension()) {
 			return core::make_shared<CubzhB64Format>();
+		} else if (ext == aseprite().mainExtension()) {
+			return core::make_shared<AsepriteFormat>();
 		} else if (ext == roomsThing().mainExtension()) {
 			return core::make_shared<ThingFormat>();
 		} else if (gltf().matchesExtension(ext)) {