+++ src/modules/voxelformat/private/aseprite/AsepriteFormat.h
@@ -1,0 +1,33 @@
+/**
+ * @file
+ */
+
+#pragma once
+
+#include "voxelformat/Format.h"
+
+namespace voxelformat {
+
+/**
+ * @brief Aseprite format
+ *
+ * https://github.com/aseprite/aseprite/blob/main/docs/ase-file-specs.md
+ *
+ * @ingroup Formats
+ */
+class AsepriteFormat : public RGBASinglePaletteFormat {
+protected:
+	bool loadGroupsRGBA(const core::String &filename, const io::ArchivePtr &archive,
+						scenegraph::SceneGraph &sceneGraph, const palette::Palette &palette,
+						const LoadContext &ctx) override;
+	bool saveGroups(const scenegraph::SceneGraph &sceneGraph, const core::String &filename,
+					const io::ArchivePtr &archive, const SaveContext &ctx) override {
+		return false;
+	}
+
+public:
+	size_t loadPalette(const core::String &filename, const io::ArchivePtr &archive, palette::Palette &palette,
+					   const LoadContext &ctx) override;
+};
+
+} // namespace voxelformat